// core/src/api/api-generated.ts
export type GetUserDataApplication200AllOfMeasurementStatusSummaryItem = {
  /** 横軸の時刻 */
  Time?: string;
  /** 測定状態の項目マスタ数字 */
  MeasurementStatus?: string;
  /** trueの時刻はバーの上に『▼』を表示する */
  Alert?: boolean;
};

/**
 * 『心拍グラフ』の表示データ
 */
export type GetUserDataApplication200AllOfHeartRateSummaryItem = {
  /** 横軸の時刻 */
  Time?: string;
  /** 心拍の数値 */
  Value?: number;
};

/**
 * 『呼吸グラフ』の表示データ
 */
export type GetUserDataApplication200AllOfRespirationRateSummaryItem = {
  /** 横軸の時刻 */
  Time?: string;
  /** 呼吸の数値 */
  Value?: number;
};